Refactor Pojo Vertex to move as much as possible to reusable trait
